+++ src/main/java/kr/hhplus/be/server/reservation/application/ReserveSeatService.java
@@ -7,7 +7,7 @@ import kr.hhplus.be.server.concert.domain.concertSchedule.ConcertScheduleReposit
 import kr.hhplus.be.server.concert.domain.seat.Seat;
 import kr.hhplus.be.server.concert.domain.seat.SeatRepository;
 import kr.hhplus.be.server.concert.domain.seat.SeatStatus;
-import kr.hhplus.be.server.domain.lock.DistributedLockRepository;
+import kr.hhplus.be.server.lock.domain.DistributedLockRepository;
 import kr.hhplus.be.server.queue.domain.QueueToken;
 import kr.hhplus.be.server.queue.domain.QueueTokenRepository;
 import kr.hhplus.be.server.reservation.domain.SeatReservation;
